fix: correct config defaults and computation parameters from Tejas findings
Based on Tejas's corrected experiments (scripts 10-14):
- MEASUREMENT_LAYER: 20 -> 32 (best layer for Gemma-3-4B-IT)
- REFUSAL_LAYERS: range(8,24) -> range(0,35) (full layer sweep)
- Add BEST_REFUSAL_POSITION=-2 (model token, not newline)
- Add DIRECTION_DTYPE=float64 (prevents bfloat16 precision loss)
- Float64 accumulation in refusal direction computation
- Left-padding in tokenizer (was unset, Tejas found this critical)
- Transcoder repo: add _affine suffix for IT models
- lazy_encoder=True in ReplacementModel loading
- Circuit-tracer URLs updated to AutoInterp fork
- Test assertions updated to match new config values

+++ src/refusal_lens/refusal_directions.py
@@ -175,7 +175,7 @@ def collect_resid_acts_multipos(
         
         for l in layers:
             resid = gather_residual_activations(model, l, input_ids)
-            pos_acts = [resid[0, p, :].float().cpu() for p in pos_indices]
+            pos_acts = [resid[0, p, :].double().cpu() for p in pos_indices]
             acts[l].append(torch.stack(pos_acts))
     
     logger.info("  Done: %d prompts.", len(prompts))
@@ -204,8 +204,8 @@ def compute_refusal_directions(
     results: dict[int, RefusalDirectionResult] = {}
 
     for l in harmful_acts:
-        h_mean = harmful_acts[l].mean(dim=(0, 1))
-        s_mean = harmless_acts[l].mean(dim=(0, 1))
+        h_mean = harmful_acts[l].double().mean(dim=(0, 1))
+        s_mean = harmless_acts[l].double().mean(dim=(0, 1))
         diff = h_mean - s_mean
         separation = diff.norm().item()
